Refactor layout analysis module
 - Extract `max_empty_rects` and associated types into `empty_rects` submodule.

 - Rename `page_layout` to `layout_analysis` to be more consistent with the
   naming of other modules that are named after their phase in the OCR pipeline
   (`detection`, `recognition`).
